Fit an curve through MTT data

Fit an curve through MTT data

Usage

mtt_fit(x, model, ...)

# Default S3 method
mtt_fit(x, model = NULL, ...)

# S3 method for class 'gp'
mtt_fit(x, model = NULL, ...)

# S3 method for class 'data.frame'
mtt_fit(x, model = NULL, ...)

# S3 method for class 'spectramax'
mtt_fit(x, conditions, model = NULL, ...)

Arguments

x

An object of class gplate::gp, data.frame, or mop::spectramax. See details.

model

A character vector of models that should be used to fit the data. If NULL, will start with a 4-parameter log-logistic model and fall back using a linear model. Options for supplying a vector include NA, "lm", and "ll". See details.

...

Unused

conditions

A named list of numerics of length 4. Contains doses per quadrant. Names are the drug. If quadrant should be omitted, supply NA. Quadrants go from left to right, top to bottom. Identical names are allowed.

Details

If a data.frame or gplate::gp is supplied, it should have columns condition, dose, nm562 and nm660

This function will first attempt to fit the data using a 4-parameter log-logistic model, then use a linear model as a fallback if fitting fails.

The model argument can be used to specify a model manually by supplying a character vector with length equal to the number of unique conditions. Valid values for the character vector include:

NA: Fitting will start with log-logistic, fallback on linear

"ll": Only (try to) fit with log-logistic

"lm": Only fit with lm